brokers/crypto/requirements.txt
async-timeout==4.0.3
boto3==1.34.4
botocore==1.34.4
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
s3transfer==0.9.0
six==1.16.0
urllib3==2.0.7
websocket-client==1.7.0

